Organize codebase into packages
