+++ data/duke.txt
@@ -1,1 +1,1 @@
-D | 1 | return book | 2022-09-10
+D | 0 | return book | 2022-09-10
