--- src/main/java/Deadline.java
@@ -1,32 +1,0 @@
-import java.time.LocalDate;
-
-public class Deadline extends Task {
-    private LocalDate by;
-
-    Deadline(String description, boolean isDone, LocalDate by) {
-        super(description, isDone);
-        this.by = by;
-    }
-
-    Deadline(String description, boolean isDone, String by) throws DukeException {
-        this(description, isDone, Parser.parseDate(by));
-    }
-
-    Deadline(String description, String by) throws DukeException {
-        this(description, false, Parser.parseDate(by));
-    }
-
-    String getFormattedDateString() {
-        return this.by.format(DATE_FORMATTER);
-    }
-
-    @Override
-    String getStorageFormat() {
-        return "D | " + super.getStorageFormat() + " | " + this.by;
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + this.getFormattedDateString() + ")";
-    }
-}
