+++ src/main/java/duke/command/AddCommand.java
@@ -1,7 +1,11 @@
+package duke.command;
+
+import duke.task.Task;
+
 public class AddCommand extends Command {
     private Task task;
 
-    AddCommand(Task task) {
+    public AddCommand(Task task) {
         this.task = task;
     }
