+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.task.TaskList;
+import duke.util.Ui;
+
 public abstract class Command {
     protected static Ui ui;
     protected static TaskList taskList;
@@ -12,6 +18,7 @@ public abstract class Command {
     }
 
     public abstract void execute() throws DukeException;
+
     public boolean isExit() {
         return this.isExit;
     };
