+++ src/main/java/duke/command/HelpCommand.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.DukeException;
+
 import java.util.ArrayList;
 
 public class HelpCommand extends Command {
