+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,10 +1,12 @@
+package duke.command;
+
 import java.util.ArrayList;
 
-public class EmptyCommand extends Command {
+public class UnknownCommand extends Command {
     @Override
     public void execute() {
         ArrayList<String> toPrint = new ArrayList<>();
-        toPrint.add("No command entered!");
+        toPrint.add("Unknown command provided!");
         String[] helpGuide = HelpCommand.getHelpGuide();
         for (String line: helpGuide) {
             toPrint.add(line);
