+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,37 @@
+package duke.task;
+
+import duke.DukeException;
+import duke.util.Parser;
+
+import java.time.LocalDate;
+
+public class Deadline extends Task {
+    private LocalDate by;
+
+    public Deadline(String description, boolean isDone, LocalDate by) {
+        super(description, isDone);
+        this.by = by;
+    }
+
+    public Deadline(String description, boolean isDone, String by) throws DukeException {
+        this(description, isDone, Parser.parseDate(by));
+    }
+
+    public Deadline(String description, String by) throws DukeException {
+        this(description, false, Parser.parseDate(by));
+    }
+
+    String getFormattedDateString() {
+        return this.by.format(DATE_FORMATTER);
+    }
+
+    @Override
+    public String getStorageFormat() {
+        return "D | " + super.getStorageFormat() + " | " + this.by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + this.getFormattedDateString() + ")";
+    }
+}
